weekly/requirements.txt
pytest
pylint
requests
python-dotenv